+++ packages/dso-toolkit/src/components/card-container/card-container.content.ts
@@ -1,6 +1,6 @@
 import { Card } from "../card/card.models";
 
-export const cardContainerContent: { cards: Card<unknown>[] } = {
+export const cardContainerContent: { cards: Omit<Card<unknown>, "dsoCardClick">[] } = {
   cards: [
     {
       label: "Begrippen uit de Omgevingswet",
@@ -28,7 +28,6 @@ export const cardContainerContent: { cards: Card<unknown>[] } = {
     {
       label: "Werkzaamheden",
       href: "#",
-      image: "images/rectangle1.png",
     },
     {
       label: "Waardelijsten",