+++ packages/dso-toolkit/src/components/plekinfo-card/plekinfo-card.demo.ts
@@ -1,0 +1,16 @@
+export const plekinfoCardDemoCss = `
+  .symboolcode {
+      background-color: #fff;
+      display: block;
+      float: left;
+      height: 20px;
+      margin-block-start: 2px;
+      width: 24px;
+      overflow: hidden;
+      position: relative;
+    }
+    
+  .symboolcode[data-symboolcode=vgz023] {
+    background-image: url(images/rectangle2.png);
+  }
+`;