+++ storybook/cypress/e2e/card.cy.ts
@@ -1,9 +1,9 @@
 describe("Card", () => {
   beforeEach(() => {
-    cy.visit("http://localhost:45000/iframe.html?id=core-card--href-and-selectable-with-button")
+    cy.visit("http://localhost:45000/iframe.html?id=core-card--with-selectable-and-button")
       .get("dso-card")
       .then(($card) => {
-        $card.on("dsoCardClicked", cy.stub().as("dsoCardClickedListener"));
+        $card.on("dsoCardClick", cy.stub().as("dsoCardClickListener"));
       });
   });
 
@@ -12,26 +12,28 @@ describe("Card", () => {
     cy.dsoCheckA11y("dso-card.hydrated");
   });
 
-  it("should call dsoCardClicked event when user clicks a non-interactive element in card", () => {
+  it("should only call dsoCardClick event when user clicks a title in heading", () => {
     cy.get("dso-card")
       .find("dso-selectable > .dso-selectable-container > .dso-selectable-input-wrapper > input")
       .focus()
       .realClick()
-      .get("@dsoCardClickedListener")
+      .get("@dsoCardClickListener")
       .should("not.have.been.called")
       .get("dso-card")
       .find(".dso-card-interaction")
       .first()
       .realClick()
-      .get("@dsoCardClickedListener")
+      .get("@dsoCardClickListener")
       .should("not.have.been.called")
       .get("dso-card")
+      .shadow()
+      .find(".dso-card-heading > a")
       .realClick()
-      .get("@dsoCardClickedListener")
+      .get("@dsoCardClickListener")
       .should("have.been.calledOnce");
   });
 
-  it("should not call dsoCardClicked on toggletip", () => {
+  it("should not call dsoCardClick on toggletip", () => {
     cy.get("dso-card")
       .find("div[slot='interactions']")
       .then(($cardInteractions) => {
@@ -42,48 +44,13 @@ describe("Card", () => {
       .get("dso-card")
       .find(".dso-card-interaction > dso-toggletip")
       .click()
-      .get("@dsoCardClickedListener")
+      .get("@dsoCardClickListener")
       .should("not.have.been.called");
-  });
-
-  it("should set isModifiedEvent when the card event is triggered with modifiers (eg. holding CTRL or right-click)", () => {
-    cy.visit("http://localhost:45000/iframe.html?id=core-card--selectable");
-  });
-
-  it("should not call events when clickable is false", () => {
-    cy.get("dso-card")
-      .invoke("attr", "clickable", "false")
-      .realClick()
-      .get("@dsoCardClickedListener")
-      .should("not.have.been.called");
-  });
-
-  it("should have correct image dimensions", () => {
-    cy.visit("http://localhost:45000/iframe.html?id=core-card--href-with-image")
-      .get("dso-card")
-      .find("img[slot='image']")
-      .should("have.css", "height", "24px")
-      .and("have.css", "width", "24px")
-      .get("dso-card")
-      .invoke("attr", "image-shape", "wide")
-      .find("img[slot='image']")
-      .should("have.css", "height", "26px")
-      .and("have.css", "width", "30px");
-  });
-
-  it("creates anchor when href is set", () => {
-    cy.visit("http://localhost:45000/iframe.html?id=core-card--static")
-      .get("dso-card")
-      .invoke("prop", "href", "#")
-      .shadow()
-      .find("a.heading-anchor")
-      .should("have.attr", "href", "#");
   });
 
   it("creates anchor to external link when href is set and mode is set to 'extern'", () => {
-    cy.visit("http://localhost:45000/iframe.html?id=core-card--static")
+    cy.visit("http://localhost:45000/iframe.html?id=core-card--default")
       .get("dso-card")
-      .invoke("prop", "href", "#")
       .invoke("prop", "mode", "extern")
       .shadow()
       .find("a.heading-anchor")
@@ -101,9 +68,8 @@ describe("Card", () => {
   });
 
   it("creates anchor to download link when href is set and mode is set to 'download'", () => {
-    cy.visit("http://localhost:45000/iframe.html?id=core-card--static")
+    cy.visit("http://localhost:45000/iframe.html?id=core-card--default")
       .get("dso-card")
-      .invoke("prop", "href", "#")
       .invoke("prop", "mode", "download")
       .shadow()
       .find("a.heading-anchor")
@@ -112,5 +78,16 @@ describe("Card", () => {
       .shadow()
       .find("svg")
       .should("have.attr", "id", "download");
+  });
+
+  it("should show different background-color when active='true'", () => {
+    cy.visit("http://localhost:45000/iframe.html?id=core-card--default")
+      .get("dso-card.hydrated")
+      .invoke("prop", "active", true)
+      .shadow()
+      .find(".dso-card-container")
+      .should("have.css", "background-color", "rgb(229, 229, 229)");
+
+    cy.get("dso-card.hydrated").matchImageSnapshot();
   });
 });