+++ storybook/src/components/document-card/document-card.content.ts
@@ -1,0 +1,6 @@
+import { html } from "lit-html";
+
+export const typeItems = [
+  html`<span class="dso-document-card-type-item">Omgevingsplan</span>`,
+  html`<span class="dso-document-card-type-item">Gemeente Amsterdam</span>`,
+];