+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Next
+### Added
+* **BREAKING** Card: Opsplitsen naar Card, Plekinfo Card en Document Card ([#2797](https://github.com/dso-toolkit/dso-toolkit/issues/2797))
 
 ### Docs
 * Voorbeeldpagina: Mijn overzicht bijwerken ([#2785](https://github.com/dso-toolkit/dso-toolkit/issues/2785))